paranoid pass: 24 asm-verified silent-bug fixes across 24 TUs
Every fix independently re-verified against the BURNOUT_X360_ARTIST.XEX
disassembly (6 adversarial verifiers, 24/24 CONFIRMED, 0 refuted). Build
links green; boot reaches EA->Criterion videos with no asserts.

Width/sign reads (load-instruction mismatches):
- CgsID CgsIDCompress: u8 -> s8 (extsb; signed char compares)
- CgsAptInterpolator GetCurrentValue: f64 -> f32 (lfs not lfd @+0x14)
- CgsModelResourceType: import count/fields @+16/+18 u16 -> u8 (lbz)
- CgsRwRenderableResourceType DebugValidate: byte offsets 18/20 + lbz @+8
  (Hex-Rays int* arithmetic misread as bytes 72/80)

Descriptor fills / layout:
- Shader parameter/parametersemantic: 5 (size,align) pairs = (size,16)+(0,1)x4
  (entries had value/alignment swapped and wrong tail)
- CgsModuleSingleBuffered Prepare: ALLOCINPUT resume stage 4 -> 2

Constants / init:
- BrnShaderConstantsFrame mfWhiteLevel 0.0 -> 1.0 (would black out scene)
- BrnLoadingScreenRenderer mfArrowDirection 0.0 -> 1.0 (would freeze bounce)
- BrnGameMainFlowController meCurrentState INVALID(-1) -> INITIAL_LOADING_SCREEN(0)
- CgsGuiState State::Construct: zero mbIsSaveLoadState/mbIsVideoState (+0x31/+0x32)
- BrnMomentParameterBank: mbClose/mbTakedown last-write-wins corrections
- BrnResourceRegistrar QueuedResource: zero-init mbResourceStillRequested (+0x140)

Logic / control flow:
- BrnGameModule GameMain: 3 perfmon brackets to correct CpuMonitor handles
  (DebugManager +0x48, Network +0x10, RenderAll +0x4C)
- BrnResourceRegistrar UpdateRequests: guard = FindFirstInstanceOf==-1
  (not refcount==0; refcount gate lives inside AddNode)
- CgsDebugManager: null path default "Debug/Components"; ActivateComponent
  no-op when already active
- CgsDebugComponent MakeFullPath: read buffer[strlen] (NUL) per asm
- CgsFsm Update: re-read mpCurrentState before each virtual (no caching)
- CgsScriptedState Construct: real sig (CgsID, ScriptedFsm*); std@+8/stw@+0x10,
  no self-ref (was a Hex-Rays 64-bit-a1 artifact misread)
- CgsMemoryModule: return 4 (no parent blocks) / 7 (size not multiple of granule)
- CgsDeviceManager OpenEx: callback result = 1st out-param (sign-extended)
- ZoneListResourceType: FixDown delta = src (not 0)
- CgsAptDataHeaderType: import-table walk indexes/derefs slot->entry, entry[1]
  test, offset from entry pointer (not slot address); 4-byte cursor stride
- CgsDebugRenderStreamReader: DoAllocate name arg NULL (fabricated labels removed)

+++ src/GameShared/GameClasses/Gui/View/AptInterface/CgsAptInterpolator.cpp
@@ -3,23 +3,27 @@
 // Reconstructed from BURNOUT_X360_ARTIST.XEX @ 0x824E5BC8
 //   (CgsGui::Interpolator::GetCurrentValue)
 //
-// Behaviour-faithful to the X360 pseudocode:
-//     return *(this + 20);          // returned as double
+// Behaviour-faithful to the X360 asm:
+//     lfs f1, 0x14(r3)  ;  blr        // load a 32-bit float at +0x14, return it
 //
-// The current interpolated value is stored at byte offset 0x14 and returned as a
-// double. (Hex-Rays types the load as a double at +0x14; modelled as such.)
+// The current interpolated value is stored at byte offset 0x14 as a 32-bit FLOAT
+// and loaded with lfs (load float single), not lfd. (FLAG: the Hex-Rays
+// pseudocode types the load as `double` at +0x14, but the ARTIST asm @0x824E5BC8
+// is `lfs` and DecFIGS @0xBB3D08 demangles to a `float32_t`-returning
+// GetCurrentValue with the same `lfs 0x14` -- so the member is an f32, loaded as
+// f32. A double member + lfd would read 8 bytes and misinterpret the bit pattern.)
 
 namespace CgsGui
 {
     struct Interpolator
     {
         u8   mPad[20];          // [0x00] opaque
-        f64  mfCurrentValue;    // [0x14] current interpolated value
+        f32  mfCurrentValue;    // [0x14] current interpolated value (32-bit float)
 
-        double GetCurrentValue() const;
+        float GetCurrentValue() const;
     };
 
-    double Interpolator::GetCurrentValue() const
+    float Interpolator::GetCurrentValue() const
     {
         return mfCurrentValue;
     }
